pyproject.toml
[tool.poetry]
name = "RitoAPI"
version = "0.1.0"
description = "A mock Riot API to test rate limiting against."
authors = ["Robin (Nurse) <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"
requests-html = "^0.10.0"
PyYAML = "^5.4.1"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
